Files not found when baseUrl is not `/`
The following assumes a default `--out-dir` of `build`.

Assume a single file: `docs/intro.md`. when building with
docusaurus 2.0, you get `build/docs/intro/index.html`

When you use `/` as the baseUrl, which is the default,
the previous code works correctly.

When you use `/some-thing/` as the base url, your docs are still
written to `build`, yet the `route` received will be:
`/some-thing/docs/intro/index.html`. This route, appened with
the outputDir will result in `build/some-thing/docs/intro/index.html`
which does not exist.

This PR changes the path to look at depending on the baseUrl. If the
default is used (`/`), then we use the route as is, otherwise, we drop
the baseUrl from the `route` via substring.

-----

To reproduce,
1. follow the quickstart guide for docusaurus.
2. install lunr
3. build (it works)
4. change the baseUrl
5. build (it fails)
6. (optional) test this PR by updathing the file under `node_modules/docusaurus-lunr-search/src/utils.js` with the content of this PR and re-run