Avoid adding extra `/` on hits
say we got doc.url = /docs/intro.html

with baseUrl = '/', the condition failed because doc.url starts with
a `/`, so we returned `doc.url

with baseUrl = `/some-thing/`, the condition succeeds because baseUrl
is not `/`, so we return the concatenated url, but it's wrong. it gives:
`/some-thing//docs/intro.html`

The thing is that the doc.url already contains the baseUrl, so there
is no need to concatenate.