Use ⌘ to trigger search bar for Mac users
The standard way to trigger a search bar on websites for Mac users has become the command key (see for example tailwind.com or github.com).

This PR allows focusing the search bar with either the `ctrlKey` or the `metakey`, so both platforms are supported. It also updates the placeholder text for MacOS users to show ⌘+K instead of Ctrl+K.